pyproject.toml
[project]
name = "testwork020"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "load-dotenv>=0.1.0",
    "pydantic-settings>=2.7.0",
    "pydantic>=2.10.4",
    "uvicorn>=0.34.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg==3.2.3",
    "celery>=5.4.0",
    "pytest>=8.3.4",
    "httpx>=0.28.1",
    "pytest-asyncio>=0.25.0",
    "redis>=5.2.1",
    "exrex>=0.12.0",
    "pytest-dependency>=0.6.0",
]
